Server/requirements.txt
flask
flask-cors
flask-bcrypt
flask-mail
python-dotenv
supabase
pyjwt
email-validator